pyproject.toml
[tool.poetry]
name = "coursework-num-7"
version = "0.1.0"
description = ""
authors = ["Makoto <makoto@mak.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.6.0"
drf-yasg = "^1.21.7"
setuptools = "^72.1.0"
redis = "^5.0.8"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
flake8 = "^7.1.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
